Update patch set 2
I would prefer that you didn't submit this.

We are getting ready to push out a big set of changes to the Android
tree, and I think merging across these sorts of cosmetic changes will
probably make our lives difficult.

Although we appreciate the effort of tidying the code, I don't think now
is quite the right time to do it.

Patch-set: 2