Create change
Patch-set: 1
Change-id: Id456e9c6b6d4d79c16678dc588f9df94d9cd4dc1
Subject: Contribution of IA-32 assembly for the fast byte-code engine. This contribution is the IA-32 assembly implementation for the fast byte-code interpreter engine. This is an initial contribution with minimal optimizations that target the Intel ATOM processor. We expect to continuously improve the IA-32 fast interpreter. This contribution is expected to elicit discussion on the potential merge of this code and similar efforts (i.e. the mterp/x86 directory included with cupcake). Using the simulator build for AOSP master we were able to show the following improvements against the c-portable interpreter:
Branch: refs/heads/master
Commit: d456e9c6b6d4d79c16678dc588f9df94d9cd4dc1