Update patch set 1
I would prefer that you didn't submit this.

I think you've got the right idea with using "expr --" instead of just
"expr" to process the -J options. I think it can be done in a more
project-neutral manner by always using "expr --". I tested this using
SLAX and it behaved the same way.

I am curious why lines 65 & 66 of the original file use different
parameters. The fact that 'x-J' is used suggests that someone
encountered this error before, but did not need to fix it on the
following line. As such, using 'x-J' is an alternative to using "expr
--" if this is found to be incompatible with certain systems.

A grep of the project also shows a similar problem in:

* build/tools/apicheck
* development/tools/traceview

Being new to Android development I'm not familiar with the submission
process, is there a way to submit a single change to all three projects?
If not I can make the change to the two other files and submit that.

Patch-set: 1