Update patch set 1
No score.

> The fact that 'x-J' is used suggests that someone
> encountered this error before, but did not need to fix it on the
> following line. As such, using 'x-J' is an alternative to using "expr
--" 
> if this is found to be incompatible with certain systems.
Actually problem occured on next line (76 in commited source), where $1
is expanded to -Jsome_vm_opt.

I've not tested this code on Linux distributions other than CentOS 4.4
(expr 5.2.1), so I've decided leave it only for FreeBSD.

> A grep of the project also shows a similar problem in:
> * build/tools/apicheck
> * development/tools/traceview
In fact, x[Something] is used not only in this shell scripts in Android,
e.g. in conditions to check if variable is not empty like: if [[ x${VAR}
!= x ]]

I don't like such syntax, but it works anyway. You may fix it, it'll be
better code, IMHO.

> Being new to Android development I'm not familiar with the submission
> process, is there a way to submit a single change to all three
projects?
> If not I can make the change to the two other files and submit that.
AFAIK, it's impossible to make change for all three of them in one
changeset, cause it's different projects and thus in fact - different
git repositories. So, just make three commits.

Patch-set: 1