Update patch set 1
I would prefer that you didn't submit this.

First and foremost, thanks for making this submission. We look forward
to getting it included. However, we have a few requests to get things in
order:

* Please change the copyright header to say "The Android Open Source
Project". 
* Please rename the directory to be "x86-atom" since (a) we use the
standard term "x86" to refer to the architecture in question and (b) it
makes it clear that it is specialized for Atom.
* We will probably have some more detailed comments later about how to
hook the code up in the build system, so please be prepared to make a
few tweaks in that stuff.

I also want to add that we ultimately would like to merge the code for
all x86 variants into a single mterp subdirectory, but keeping your work
in a separate area is a reasonable short-term step.

Thanks again!

Patch-set: 1