Update patch set 1
No score.

This is to address concerns Dan Bornstein had raised with change 6426.
We originally wanted to use 'expr --' to avoid having -J operands appear
as options. However, it is still possible that the operand and operator
will get confused. It is better to precede operands with 'x', as on the
previous line, as this will always work to distinguish between the two.

Patch-set: 1