Update patch set 1
No score.

Thanks for the comments. We are fine to put the code in x86-atom or
x86/atom directory under mterp. Our code is mostly generic x86 at this
version. We plan to gradually merge our generic x86 implementation with
existing code under the x86 directory and keep atom specific code under
x86-atom.

We have checked with Intel legal regarding the copyright header. The
corporate contributor's license we did with AOSP is a license to our
copyrights, not an assignment.  Hence our legal insists that we keep the
copyright header that says “Intel Corp.” Please clarify the request to
change copyright header to "The Android Open Source Project" since it
does not conform to the CCL we have signed.

Thanks.

Patch-set: 1