Update patch set 1
I would prefer that you didn't submit this.

I am not qualified to make a detailed argument about the copyright line.
However, my understanding is that the standard "Android Open Source
Project" line is sufficient for all the usual protections.

What I *do* know is that the code you are contributing is just one
directory in a sea of "Copyright The Android Open Source Project," where
most of that so-copyrighted code actually came from Google, though
Google is not given any credit in the copyright notice. Taking this
patch is in part a statement that we want this code as part of The
Android Open Source Project, and that, in at least some regard, the
project as a whole is taking responsibility for it. So, it would seem
weird to me if this one piece had to have a special copyright for
itself.

If you are concerned about losing credit, then please just add something
like "The Atom-specific version of the interpreter core was contributed
by Intel." to the top-level README.txt file.

If this is insufficient, then I think there needs to be a non-technical
discussion, which is best taken up in a different forum.

Just by way of analogy, as we contribute patches from our core library
back to the Apache Harmony project, we fully expect to see the copyright
headers of those files replaced with the standard Harmony header.

Thanks again, and warm regards in the new year.

Patch-set: 1