Update patch set 1
Patch Set 1: No score

Hello, thanks for your comments and your fast response. We will convert the tabs to spaces when we resubmit the patch. 

It is our teams procedure to run a series of framework, vm validity tests, and benchmarks before submitting a patch. Using a modify emulator build we are able to pass all but the 044-proxy test. That failure is unrelated to this patch. Using the simulator build with the command "./run-all-tests --local --fast", we are able to pass all but the following tests (whose failures are also unrelated to this patch).

passed: 66 test(s)
failed: 5 test(s)
failed: 030-bad-finalizer
failed: 044-proxy
failed: 057-iteration-performance
failed: 062-character-encodings
failed: 063-process-manager

Andy, did you try to run the code and not get the same result, or did you just want to make sure we passed these tests before we submitted the patch?

Thanks.

Patch-set: 1
Reviewer: Johnnie L Birch Jr <1000706@85c56323-6fa9-3386-8a01-6480fb634889>