Update patch set 1
Patch Set 1: No score

I can offer two alternatives.

Alt-1 is very short in code (10 lines all in all), it makes use of an 8 or 9 bit hash, does therefor require a fixed sized hash array, does not involve any looping so there is practically no CPU overhead. But because of possible hash collisions, it is not 100% efficient. Some small amount of garbage will be created still. (Less so with the 9 bit hash.) It's still much better performing than the current solution. The trade-in in this case is the hash array of 256 or 512 elements. The solution involves very simple to understand code.

Alt-2 combines Alt-1 with my original approach, it uses an 8 bit hash array together with a small overflow array for hash collisions. For most apps zero garbage is being created. Animations are really smooth. But in case of hash collisions a small amount of looping will take place. This solution involves more code, which I see as a disadvantage.

Would you like to review any of these?

Patch-set: 1
Reviewer: Timur Mehrvarz <1001563@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0