Initial changes to support counting mark/sweep by generation.
Tallies, by bytes and object counts, the difference between the current GC
number and the GC when this object was created
Includes a way to force GC before bumping a failed alloc, so "generations"
are all roughly the same size, where "size" means number of total bytes
allocted. This creates poor-man's generations.
Somewhat bogus in the presence of user-forced GCs or GCs which get triggered by
a single object larger than the number of bytes which will force a GC.