Update patch set 1
Patch Set 1: No score

I don't know the "deep answer" to "why does the zygote require this". I do know that all the things I saw failing the orig test were (1) in zygote and (2) had a scanGeneration of 0. I had a comment, but it seemed to be just rendering the code into English, e.g:
Don't abort if this is the zygote and scanGeneration is zero.
because that' pretty much where my knowledge ends.

I haven't hunted for the place where [in the zygote, I assume] objects with scanGeneration == 0 get created. Is that what you're asking for?

In some ways it might be better to fix this problem by fixing that place to patch the scanGeneration. Would you like me to explore that as well?

Patch-set: 1