Create change
Patch-set: 1
Change-id: I3ca3a25d710654bfd10643794bb5aba1b2792065
Subject: Move the code which sets the scanGeneration [and other fields] of a chunk in scanObject to before the early return in the cases where clazz is either NULL or unlinkedJavaLangClass. This way the "wasn't scanned last time" check in _markObjectNonNullCommon will see the scanGeneration and markGeneration as equal, and won't dvmAbort.
Branch: refs/heads/master
Commit: 3ca3a25d710654bfd10643794bb5aba1b2792065