Update patch set 1
Patch Set 1: No score

Two thoughts:

(1) Check the spacing on the CallEABI.S -- want a space between operands, and no space at the end of the line (shows up in red on the side-by-side diff).

(2) This introduces TARGET_ARCH_VERSION to the build system, which means this goes beyond a simple Dalvik change.

I'm happy with the CallEABI.S change (other than the minor formatting nits), but not quite comfortable with the makefile change -- need buy-in from the build folks.  Might make sense to do the makefile alteration in a different patch.

Patch-set: 1