Update patch set 1
Patch Set 1: No score

Without change, booting emulator and waiting for logcat to settle:

  PID      Vss      Rss      Pss      Uss  cmdline
  568   24832K   24232K   10716K    7236K  system_server
  609   20620K   20620K    8808K    6228K  android.process.acore
  607   17328K   17328K    6791K    5016K  com.android.phone
  541   19788K   19788K    6747K    3668K  zygote

With change:

  PID      Vss      Rss      Pss      Uss  cmdline
  569   24888K   24288K   10773K    7288K  system_server
  609   20632K   20632K    8824K    6240K  android.process.acore
  610   17428K   17428K    6892K    5112K  com.android.phone
  541   19788K   19788K    6734K    3660K  zygote

I'm not seeing a meaningful change in the USS column.  If this were improving sharing by 100K there should be a noticeable difference.

It's possible my git-fu is weak, but you may want to test with procrank on your end and see if you're getting the same thing.

Patch-set: 1