Update patch set 1
Patch Set 1: No score

I'm on the fence about this change. A very quick test seems to show that the difference between the copy and the live heap doesn't account for the private COW pages, which is what we really want to track.

This makes sense if there are A->B->A writes, or even A->A, like storing zero or null where it's already zero or null. That copies the page, but doesn't show in the diff.

The really right answer, I think, is using ptrace to trap actual writes. I'm trying to be brave enough to do that.

If we *had* a word-by-word dirty table, would this code be useful for anything else?

Patch-set: 1
Reviewer: Barry Hayes <1001707@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0