Update patch set 1
Barry said:
Hmm.
I think ANY change that's going to track "why are we writing stuff in this mspace" is going to have some changes to mspace.c. Or be yuckier and have all the code that "should" be in mspace live somewhere else.

This CL is "copy this mspace". if we ptraced there will be some code to deal with that instead, like it'll need a place to stash the dirty-word bits of the mspace. Or live in some off-mspace yucky table. So I'm not sure you have selected a very good reason to object to this change from among the many good reasons to object to this change ;-)

===

dbort says:

The mspace thing is just the final nail in the coffin; the fact that this isn't immediately useful and that not even you are convinced of its worth are the overriding reasons.  Yes, if we find a debugging method that actually helps with this problem, we may need to touch other pieces, but until that time comes I'd rather avoid adding a bunch of dead code.

Patch-set: 1