Update patch set 1
Patch Set 1: No score

This is a reworking of 
https://review.source.android.com/10482
The above CL has a defect:

Briefly: 1) refOffsets is initialized to zero. 2) <clinit> is called 3) <clinit> creates an instance of this class, populating ref fields of the instance. 4) GC happens, and sees 0 in refOffsets, and thus does NOT mark the ref fields 5) refOffsets gets calculated [correctly] 6) initialization ends

The object created in (3) now has pointers to unallocated memory.

This new CL has "belt and suspenders" to fix the defect:
1) The refOffsets field is initialized to CLASS_WALK_SUPER rather than zero. This should cause the chain to be walked, rather than the GC believeing that the object in (3) has no references.
2) The code to correctly populate refOffsets has been moved to before the call to <clinit>, so user code will see a correctly set refOffsets value.

Patch-set: 1