Update patch set 1
Patch Set 1: No score; I would prefer that you didn't submit this

I think class initialization is too late.  This falls into the "class preparation" category.

Let's move this into a function that takes clazz as an argument, and call it from dvmLinkClass() after the precacheReferenceOffsets stuff happens.  That's the last point where we fiddle with the instance fields.

Patch-set: 1