Update patch set 1
Sure, that sounds good.

Also, the "suspenders" of setting
        clazz->refOffsets = CLASS_WALK_SUPER;
The problem is that zero is a pretty common value -- no ref fields -- and so a bad marker for "can't encode".

It's enough that we can be sure "GC will not see an object of this class until refOffsets has been correctly initialized.
But having proven that I know many wrong and buggy ways to write this code, I would like to make sure that the field is initialized as above, too.

There's an "obvious" place in findClassNoInit. Are there other places where classes get created where I should make sure the field is set correctly?

Patch-set: 1