Update patch set 1
Patch Set 1: No score

I did a quick test with desktop Java and confirmed that substituting "_00024" in for "$" is the expected behavior.  Which is a little weird, since "javah" seems to simply replace '$' with '_'.

The only thing that might be worth doing differently here is changing the comparison to use isalnum().  I don't think anything other than A-Za-z_;[/$ can legally appear in a Java "binary" name, and we're catching the others, so it's probably not necessary.

Patch-set: 1