Update patch set 1
Patch Set 1: No score; I would prefer that you didn't submit this

The way this is solved in the rest of the code is to prefix strings that might start with "-" with "x". This is a practice that is explicitly recommended in the POSIX spec, IIRC. And there is an example of it one line up, even!

Patch-set: 1
Reviewer: Dan Bornstein <1000369@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1
Label: Verified=0