Create change
Patch-set: 1
Change-id: I1bc62f13ee0180b01b111b979dca28b9347cf2d6
Subject: There is multiple bugs when trying to connect a SocketChannel registered in a Selector: - The selector will generate spurious socket-connect ready   operations.  Calling finishConnect on this ready op will   trigger NoConnectionPendingException. - If the OP_CONNECT operation is not removed from the interest   set then subsequent call to select() or select(long) will return   immediately and the thread will probably 100% of the CPU. The problem is that the underlying file description is registered in the OS select() as soon OP_CONNECT is requested when it should only be set when the connection is pending.
Branch: refs/heads/master
Commit: 1bc62f13ee0180b01b111b979dca28b9347cf2d6