Add attempts to GC before failing external allocations.
We were experience the same symptoms as described in http://code.google.com/p/android/issues/detail?id=2426
and here is a patch that solves the problems for us. The discussion on the bug mentions
that such a solution would not work due to the lack of a compacting GC but this turns
out not to be the case for us. The reason is that collecting a few small bitmap objects
really makes a huge difference for the external allocations.

Calling the GC prior failing allocations is really the normal way for the VM
to act so and managing external allocations should not be different.