Update patch set 4
Patch Set 4: No score; No score

I think the generation script gets bent out of shape if the expected and supplied set of opcodes don't match.  So you'll need to add it to this checkin.  (Alternatively, leave x86-atom out of rebuild.sh so it isn't actually generated.)

You can use a C stub implementation for now, e.g.

  op OP_THROW_VERIFICATION_ERROR c

in the config file.

We'll need to do something similar for OP_BREAKPOINT and OP_EXECUTE_INLINE_RANGE on our end when we merge to tip-of-tree.

Patch-set: 4
Label: Code-Review=0
Label: Verified=0