Update patch set 2
Patch Set 2: No score; No score

Calling runFinalizationSync is a bit dangerous.  If the thread making an external allocation holds a lock, and a finalizer synchronizes on that lock, the VM will deadlock.  Apps can get away with calling it explicitly because they can call it when they know it won't be a problem, but baking it into the VM as a side-effect of allocating memory is risky.

Of course, you don't necessarily need to run *all* finalizers here, just the ones associated with objects that will release external allocations.

Patch-set: 2
Reviewer: Andy McFadden <1000483@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0
Label: Verified=0