Update patch set 2
Ok, that doesn't sound good. Waiting for only finalization of the objects that could free up external allocations would of course be best, but that seems to require quite a bit of change. We could take a look at that, unless you already have an idea of how that could be done. 
Until that happens, could it be acceptable to replace the call to runFinalisationSync with something like:

   dvmChangeStatus(NULL, THREAD_VMWAIT);
   dvmLockMutex(&gDvm.heapWorkerLock);
    /* Wake up the heap worker and wait 10 ms for it to finish. */
    dvmSignalHeapWorker(false);
    pthread_cond_timeout_np(&gDvm.heapWorkerIdleCond, &gDvm.heapWorkerLock, 10);
    dvmUnlockMutex(&gDvm.heapWorkerLock);
    dvmChangeStatus(NULL, THREAD_RUNNING);

Patch-set: 2