Update patch set 2
You could do something like that, though it has a lower level of determinism that I'd like.  Depending upon how busy the system is and how much stuff has accumulated in the heap worker queue, X.Y seconds may not be long enough.  Sometimes it'll work and sometimes it'll fail.

OTOH, since this is essentially a hail-mary before throwing an OOM, maybe it's good enough.

The real fixes here are either (a) use Bitmap.recycle() to explicitly manage the storage in the app, or (b) teach Bitmap that expecting the VM to manage non-VM storage is a bad idea.

Patch-set: 2