Update patch set 7
Better fix is to have build/core/combo/linux-x86.mk define a default arch variant.  (I have an internal change pending for that.)

I don't know if it's better to try to push the build change out, or just check this in as-is and remove the unnecessary bit later.  If the latter, we should (a) insert a "remove me later" TODO comment, and (b) get rid of the trailing whitespace on the newly added code.

Patch-set: 7