Update patch set 1
Patch Set 1: No score; Do not submit

This change is unnecessary and incomplete.

It is incomplete because, for example, if vsrc=2 then only arg0 and arg1 are initialized.  arg2 and arg3 are not.  So you only correct the initialization for the vsrc=0 case.

It is unnecessary because the argument values are not actually used.  If vsrc=0, that means none of arg0,1,2,3 will be used by the called function.  The compiler can't know this and complains, but a more sophisticated program like valgrind will correctly determine that the uninitialized values aren't used.

We actually did fix this in the tip-of-tree sources, by uncommenting the bulk initialization (arg0=arg1=arg2=arg3=0).  This was done solely to eliminate the compiler warning when building the portable/debug interpreter.

Patch-set: 1
Reviewer: Andy McFadden <1000483@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-2
Label: Verified=0