Update patch set 1
Patch Set 1: No score; I would prefer that you didn't submit this

We've already done the same thing, in a different way, for an upcoming release of the software.  In our case it's always enabled rather than only when the monitor tracking define is enabled.

The trick we're using is, when dumping the stack trace, we check to see if the thread is in the MONITOR state.  If so, we decode the current instruction in the top stack frame, which must be a monitor-enter (unless somebody called MonitorEnter from JNI, which our code currently never does).  From that we can determine the register, pull the object reference off the stack, examine its lock, and figure out which thread currently holds it.

Then, over in the stack dump in interp/Stack.c, we print a line identifying the monitor we're waiting on and the thread that currently holds it.

It's a bit of an adventure because we have to examine the status of another thread, but we can validate objects and stack references before chasing them, so catching it at the wrong time won't cause a crash.

As you point out, this is very helpful during ANR analysis because it makes it easy to figure out what something is hung up on.  The set of monitors held by a thread can be determined statically from the source code and stack trace, but that's a pain to do by hand.

Patch-set: 1
Reviewer: Andy McFadden <1000483@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1
Label: Verified=0