Update patch set 1
Patch Set 1: No score; No score

http://graphics.stanford.edu/~seander/bithacks.html#CountBitsSetParallel

gcc wants a small change to the "& 0xF0F0F0F" part:

  warning: suggest parentheses around '+' in operand of '&'

This is, of course, a slightly silly thing to optimize given how it's used.  The new form is 2x the size, but we're talking about ~32 bytes, so I'm not so concerned.

Patch-set: 1
Label: Code-Review=0
Label: Verified=0