Update patch set 2
Patch Set 2: No score; No score

I have to pull the patch, verify that it compiles, and then exercise the code in various ways to make sure it functions as well as it did before.  This takes time that could be spent on other things.

A common saying is, "if it ain't broke, don't fix it".

However, I did kind of ask for this with the "TODO" comment, and it's a pretty simple piece, so I'll take care of it.

Patch-set: 2