Update patch set 1
Patch Set 1: No score

In Java specification of tanh there is the statement: "If the argument is zero, then the result is a zero with the same sign as the argument."
In CTS 1.5 this condition was tested as part of the tanh tests, and that`s OK.
In CTS 2.1r1 the added "KnownFailure" annotation allows that platforms that do not follow this condition receive a final "PASS" in the test.
On the other hand, platforms that do not fail any test receive a final "FAIL", because the expected failure does not happen.
If the idea is not relax the test, than the "KnownFailure" annotation should be removed so that the test case is in agreement with Java specification.

Patch-set: 1
Reviewer: LUIZ FERNANDO STEIN WETZEL <1003604@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0