Update patch set 1
Patch Set 1: No score; No score

The problem with the change is that it may not do anything at all.  javac will emit code for it, but basic optimization code will realize that the assignment is unnecessary and remove it.  I suspect that the only reason it's having an effect here is because we have debugging info turned on and "dx" doesn't want to confuse the debugger by deleting the line.

Should the code get JITed, that assignment will disappear.

The type-precise GC knows the difference between uninitialized and initialized registers, but only for first use.  In this case it's in a loop, so what really matters is knowing when a register *stops* being useful.  We need live-precise GC for that.

It would be better all around to not depend on the behavior of finalizers, which are not guaranteed to run at all ever.

Patch-set: 1
Reviewer: Andy McFadden <1000483@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0
Label: Verified=0