Update patch set 1
Patch Set 1:

Good point about the compiler optimization.  This happened to fix the problem for me, but it makes sense that this is iffy at best.

I understand that the real problem is that finalizers are evil, but I'm not exactly in a position to rewrite the Timer class at that high a level.  Is there some way I can correct the existing code to get around the invisible refs in a more reliable way that isn't going to be undone by the compiler?

Patch-set: 1