Update patch set 1
Patch Set 1:

Yes, explicitly calling Timer.cancel() will cause the Timer to clean up correctly.  In most places, our code is doing that anyway.  I just came across this problem because in a couple places we aren't, and there's nothing in the documentation that says you have to, so I tried to find a way to fix the problem.

It's admittedly a pretty small edge case, so I don't really feel like I need to scorch the earth correcting it if there isn't an easy solution.  But since this leak can appear in code which is doing everything correctly according to the documentation, I was hoping to find some way to fix it.  It sounds like my patch won't do the trick, though.

Patch-set: 1