Update patch set 1
Patch Set 1:

Ok, well if the right solution is that everybody should just be calling cancel() on their timers, then I can just fix our local code to do that, and all will be well.  I just wasn't sure whether the docs were wrong, or the code was.

Before closing this out though, does anybody see any easy way that this could, in fact, be corrected?  It's fine to update the docs to just say not to do this, but if there's some relatively painless way to fix the problem, that would probably be nicer still.  It sounds like a wiser GC would be able to deal with this, but barring that, is there any way to reliably nix the invisible reference in that variable?

Patch-set: 1