Update patch set 1
Patch Set 1:

The problem is that the invisible reference is... invisible. Any change we make will be fragile because it logically does nothing, and so both tools and developers would want to remove it!

Another potential "fix" for this problem is to null out the reference from the Task to its Timer. Is that feasible?

Patch-set: 1