Update patch set 1
Patch Set 1: Fails

We've moved libcore to a separate project, and this change can't merge as is as it targets a file that doesn't exist any more.

If this change is still relevant, please create an equivalent change in the libcore project.

Thanks for your contribution, and sorry about the annoyance.

Patch-set: 1
Reviewer: Jean-Baptiste Queru <1000660@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=-1