Update patch set 1
Patch Set 1:

I've been working with Johan offline to reproduce the problem. He was able to reproduce the problem and confirm that for my test case, taking out this lock does not reintroduce the crash. However, I'm not sure my test which focused on the original native crash issues would particularly excercise the finalizer lock case, since the sessions are either reused or probablty not cached in the native code.

Patch-set: 1