Update patch set 1
Patch Set 1: No score; Do not submit

As digit noted, Dalvik needs to be portable to systems that don't have pthread_setname_np().  In particular, this would break the sim-eng build.

For this to be useful, we'd likely want a HAVE_PTHREAD_SETNAME_NP in AndroidConfig.h.

Patch-set: 1
Label: Code-Review=-2
Label: Verified=0