Update patch set 1
Patch Set 1:

It seems to me that the rest of the file was mostly formatted with 80 columns.

Other than that, the code merges and builds fine both in the open-source tree and in Google's internal master tree. Reviewers, if you're OK with the 100-column-ness, please mark this change as verified with no further worry.

Patch-set: 1
Reviewer: Jean-Baptiste Queru <1000660@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0
Label: Verified=0