Update patch set 1
Patch Set 1: Do not submit

The VM needs to run on systems that don't have the Android-specific system property infrastructure built into initd.  For this reason, all system property parsing is done elsewhere (e.g. in AndroidRuntime.cpp in frameworks/base/...) and the results are communicated into the VM via command-line options.

To avoid having to edit lots of code for trivial debugging stuff, we recently added support for "dalvik.vm.extra-opts", which allows you to pass arbitrary arguments into the VM through the framework.

As for the rest of the change, cshapiro was recently (this week!) adding a very similar feature in a different way as part of reworking the hprof implementation.  He can provide more information about that.  (I do like that yours only fires on *uncaught* OOM, though it may be a little confusing  since the failing thread will have unrolled and lost any thread stack roots.  If that thread was holding a lot of memory, you'll see a lot of stuff in the dump that looks like it could be collected.)

Patch-set: 1
Reviewer: Andy McFadden <1000483@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-2
Label: Verified=0