Update patch set 1
Patch Set 1:

This is a well motivated change.  Can you tell me more about the allocation profile of an application that was able to benefit from this change?  Did you try any another approaches to solving this issue?

I am somewhat partial to making a more general fix.  For example, if we are able to run the finalizer at a higher priority, this would help release other resources, such as byte buffers, before otherwise causing an OOM.

As an aside, in GB we moved some view caches to a scheme which aggressively recycles bitmap objects lowering the frequency of OOMs caused bitmap allocations.  Recently, we  moved the pixel data to the heap and obsoleted the external allocation tracking facility.

Patch-set: 1