Update patch set 1
Patch Set 1:

There are a number of apps (our own and 3:rd parties) that benefit from this change. There are things that can be done in the apps, but not very pretty ones (e.g. run gc and sleep for a while and then retry bitmap allocations). We did this https://review.source.android.com/#change,13252 in an earlier project.

If external allocations are going away altogether then this change might be obsoleted at that point.

Patch-set: 1