Update patch set 1
Patch Set 1:

Elliot, I saw that the change you referenced was for OldNumberFormatTest, but this change is for NumberFormatTest. Is there a reason why we can't go ahead and submit this?

The other change is in master, so it will not be part of Froyo or Gingerbread CTS.

Patch-set: 1
Reviewer: Brian Muramatsu <1003566@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0