Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

This needs to be approved in FroYo. The bug is fixed in master, which means it won't appear until Honeycomb or later, but this is still an issue on FroYo devices. The merge will have to be dealt with.

For some reason I don't have +2 on this component, so somebody needs to +2 it...

Patch-set: 1
Reviewer: Dan Morrill <1000962@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1
Label: Verified=0