Update patch set 1
Patch Set 1:

fine by me, but morrild's the one who gets to decide.

background: there was a bug in the prebuilt GCC we shipped with eclair. this test is correct, and only ever failed when built with a broken compiler (for a target that supported the instructions it was incorrectly optimizing).

Patch-set: 1