Update patch set 1
Patch Set 1: Do not submit

I'm not clear on what this change is for. This re-enables a test that was disabled in Eclair. It was disabled in Eclair because Eclair has a bug that was (evidently) fixed in FroYo.

So you're asking for a CTS waiver for a test you don't need to pass? That's great, but I'm confused why this change was proposed.

Either way we cannot accept this test, because this would require all future Eclair devices to pass it.

Patch-set: 1
Reviewer: Dan Morrill <1000962@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-2
Label: Verified=0