Update patch set 1
Patch Set 1:

We do not have good data on which third-party applications currently have problems with this, as we have had this change in place since 1.6, and reenabled it on GB as soon as we saw the problem remained. There is some discussion, including sample code demonstrating the problem, in http://code.google.com/p/android/issues/detail?id=8488
The sample application lives much longer with this change, but I was able to get the first version to crash eventually.
The problem was easy to reproduce in Google Maps about a year ago by zooming in and out a few times, but not in the current version. We first saw the problem in an application using an image cache based on soft references.

Patch-set: 1