Fix for gray object scanning in GC
Sometimes objects on the Java heap were beeing overwritten.
This happened because the garbage collection algorithm sometimes
failed to mark object that should not be garbage collected.
These objects were then deleted and overwritten, and that often
led to crashes later on.

The reason why the objects were not marked was that the last
card on the heap was never scanned for gray objects. So when an
object was located in the very end of the heap, it was sometimes
skipped while doing a concurrent GC.

This fix includes the last card (limitcard) in the scan.

Change-Id: I9128b1998a675f509c15828b6a3d99215519e6af