Update patch set 1
Patch Set 1:

Ah, I see. given that, I'd just leave it in the old filename, so that it stays in CTS. In internal master, the new location should be in CTS now.

Even though it does not automerge, this fix does need to get into other open source trees if that is where people are getting CTS tests from, since the certificate expiration problem should be happening everywhere. (which is in fact why I had noticed and fixed this internally)

Patch-set: 1