to optimize searching finalize method during classloading.
When dalvikvm loads a class, it calls the loadClassFromDex0
function in Class.c. The loadMethodFromDex function then calls
the loadMethodFromDex function in a loop to load every direct
method of this class. The loadMethodFromDex function will use
dvmCompareNameDescriptorAndMethod, which is implemented by using
strcmp to compare method name, to check if the current method is
the finalize method defined by java, and set the
CLASS_ISFINALIZABLE flag if true. The problem is that after
dalvikvm has found the finalize method in a class, it still calls
strcmp to check the following methods, which is unnecessary. This
patch simply optimized it by checking the CLASS_ISFINALIZABLE flag
before calling strcmp, so that the classloading is slightly faster.

Change-Id: Ie1ea33ce3f46c9a45605b7be6ec73622e1b0d301
Signed-off-by: xiaoxi.meng <xiaoxi.meng@gmail.com>